Solved tunnelling problem
The play state dispatcher (which includes all time-sensitive systems such as collision checking) now runs in the fixed update instead of the regular update. This means it will run at a steady 60FPS, no matter how low the display rate gets.

Also changed the way jumping and falling are calculated. Instead of keeping track of the start-time of the jump / fall, it now just keeps track of the time elapsed since the start of the jump / fall.