Bugfix: verify cursor-position validity
On the adventure and level select screen, if the last known cursor position is no longer valid, (for example because the adventure has changed since the game was last played) the cursor used to be stuck in no-mans-land, outside any nodes or roads. It now checks for this circumstance and falls back to the default position (0,0) if the position is not valid.