Users can now override the default audio settings
This means you can change the audio settings (disable the music, for example) without creating git diffs that you might accidentally commit.

Default settings will be in `assets/config/default_settings/`, they are in source control. User override files are in `.userdata/settings/`, they are .gitignored. When loading a settings file, it will check for a user file first, then if that fails, check for the default settings file.