Add temporary fix to key display arrangement
For now, keys that have not yet been collected are displayed on the exit door. This commit fixes the pattern in which they are displayed, it is now more visually appealing. In the future, this code will be rewritten. Once we know what the final door and keys will look like. (There might not be a key display on the door at all)