Add ConfigureEditorSystem and SignalEdgeDetector
Got started on the ConfigureEditorSystem, which is to be responsible for changing the brush and toggling some options. In order to do this, created a new resource: SignalEdgeDetector. This thing makes it easy to use the input handler to not only detect when a key is being pressed, but exactly when the key starts and stops being pressed.