Got started on the system that will refresh the preview
Also renamed some system files (audio -> play_sfx & preview -> preview_animation).

The RefreshPreviewSystem will listen to an event channel and will, if necessary, recreate the editor preview. That is: give existing tiles that are about to be deleted a red tinge, draw ghostly outlines for all the tiles that are about to be placed.