Improved placing tiles and ghost tiles
When placing tiles, it now does a dry-run first where it calculates which tiles will be added and which tiles will be removed. When placing ghost-tiles (ie previews) it then uses that dry-run to determine which tiles to tint red (because they're due to be removed) and where to place ghost tiles.

Also: it now no longer places tiles when the area is obstructed:
- If there are existing tiles and force_place is not on.
- If the tile that is to be placed goes outside the world bounds.

That means that this commit introduces full support for the force_place flag.